examples/gpt_gemini_showdown/requirements.txt
-e ../..
python-dotenv
quart
uvicorn
openai
tiktoken
langchain-openai
google-cloud-speech
google-cloud-texttospeech
google-cloud-aiplatform
langchain-google-genai


